fixed threshold value error

+++ VR_tracker_9.21.2020 game flexion+levels update 2017 ver/VRtracker_upper_broken/Assets/LSL2Unity/LSLController.cs
@@ -34,7 +34,7 @@ public class LSLController : MonoBehaviour
     //public Vector3 newPos;
 
     // max value before slider changes to red
-    public int colorThreshold = 0.8;
+    public float colorThreshold = 0.8f;
 
     // will return that if nothing comes
     private const float defaultValue = 0;
